feat: Fix bug in ReviewController's average rating calculation
The code change fixes a bug in the ReviewController where the average rating calculation was incorrect. The previous implementation assumed that `avgRating` was a number, but it was actually an array. This caused the new average rating to be set to an array instead of a number. The code has been updated to correctly set the average rating to the `avgRating` array.

Related to #4