Squashed commit of the following
commit 0f2e2fc2 
Author: dataforms <dataforms> 
Date: Thu Jul 04 2024 17:31:37 GMT+0530 (India Standard Time) 

    Conflicts resolved in the console.

commit e9f749cd 
Author: Rahul Ranjan <devrahulranjan@gmail.com> 
Date: Thu Jul 04 2024 14:59:22 GMT+0530 (India Standard Time) 

    feat: Add pagination to dumpProducts fixture

The code changes in this commit add pagination functionality to the `dumpProducts` fixture. The `limit` and `skip` variables are introduced to control the number of products fetched from the dummyjson API. The `dummyjsonURL` is updated to include the `limit` and `skip` parameters. This enhancement allows for fetching a specific range of products, improving the flexibility of the fixture.

Related to #5


commit 528977a4 
Author: Rahul Ranjan <devrahulranjan@gmail.com> 
Date: Thu Jul 04 2024 14:59:03 GMT+0530 (India Standard Time) 

    feat: Generate vector embeddings for product descriptions

The code changes in this commit modify the `productController.ts` file to generate vector embeddings for product descriptions. The `generate_vectors` function from the `generateEmbeddings.ts` file is imported and used to generate the embeddings. The embeddings are then added to the product data before creating a new product. This enhancement improves the search functionality by allowing users to search for products based on their descriptions.

Related to #5


commit 015e4820 
Author: Rahul Ranjan <devrahulranjan@gmail.com> 
Date: Thu Jul 04 2024 12:27:42 GMT+0530 (India Standard Time) 

    chore: Add Atlas Vector Search

Related to #5


commit a6a5d4b4 
Author: Rahul Ranjan <devrahulranjan@gmail.com> 
Date: Wed Jul 03 2024 19:33:41 GMT+0530 (India Standard Time) 

    chore #5: Add Atlas Vector Search

- Config Atlas Vector Search
- Run Cron Job
- Added Python


commit 2dc44390 
Author: Rahul Ranjan <devrahulranjan@gmail.com> 
Date: Wed Jul 03 2024 14:40:26 GMT+0530 (India Standard Time) 

    feat: Add seed and delete dump scripts for fixtures

Related to #5