feat: Add healthchecks to docker-compose services
This commit adds healthchecks to the `app-server` and `nginx-server` services in the `docker-compose-certbot.yml` file. The healthchecks use the `curl` command to test the availability of the services by making HTTP requests to the `/health` endpoint. The healthchecks are configured to run every 45 seconds with a timeout of 5 seconds and 5 retries.

Related to #5